+++ dwds/test/fixtures/fakes.dart
@@ -40,9 +40,8 @@ class FakeInspector extends Domain implements AppInspector {
   @override
   Future<ScriptList> getScripts(String isolateId) => null;
   @override
-  Future<ScriptRef> scriptRefFor(String uri) => null;
-  @override
-  Future<List<ScriptRef>> scriptRefs(String isolateId) => null;
+  Future<ScriptRef> scriptRefFor(String uri) =>
+      Future.value(ScriptRef(id: 'fake', uri: 'fake://uri'));
   @override
   Future<ScriptRef> scriptWithId(String scriptId) => null;
   @override
@@ -176,9 +175,7 @@ class FakeWebkitDebugger implements WebkitDebugger {
   Stream<ExceptionThrownEvent> get onExceptionThrown => null;
 
   @override
-  void close() {
-    return;
-  }
+  void close() {}
 
   @override
   Stream<WipConnection> get onClose => null;